# docs/spec/openapi.yaml
openapi: 3.0.3
info:
  title: authgate API
  description: |
    Central authentication service for first-party applications.
    Supports web login via Google OIDC and CLI device authorization flow.
  version: 0.1.0
  contact:
    name: authgate Team
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /oauth/authorize:
    get:
      summary: Authorization Endpoint
      description: |
        Initiates the OAuth2 authorization flow. Redirects to Google OIDC for authentication.
        Supports PKCE (required for public clients).
      operationId: authorize
      tags:
        - OAuth2
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be 'code' for authorization code flow
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: Registered client ID
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Must match registered redirect URI
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: "openid profile email"
          description: Space-separated list of scopes
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 16
          description: Opaque value for CSRF protection
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge (S256)
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
          description: PKCE method (must be S256)
        - name: nonce
          in: query
          required: false
          schema:
            type: string
          description: Optional nonce for ID token validation
      responses:
        '302':
          description: Redirect to Google OIDC or login page
          headers:
            Location:
              schema:
                type: string
              description: Redirect URL
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /oauth/callback:
    get:
      summary: OAuth Callback
      description: Handles the callback from Google OIDC after user authentication.
      operationId: callback
      tags:
        - OAuth2
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: Authorization code from IdP
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: State parameter from original request
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Error code if authentication failed
        - name: error_description
          in: query
          required: false
          schema:
            type: string
          description: Human-readable error description
      responses:
        '302':
          description: Redirect to client redirect_uri with code or error
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /oauth/token:
    post:
      summary: Token Endpoint
      description: |
        Exchange authorization code for tokens, refresh access tokens,
        or exchange device code for tokens.
      operationId: token
      tags:
        - OAuth2
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                authorization_code:
                  summary: Authorization Code Exchange
                  value:
                    access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "Bearer"
                    expires_in: 900
                    refresh_token: "opaque_refresh_token_string"
                    id_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    scope: "openid profile email"
                refresh_token:
                  summary: Refresh Token Exchange
                  value:
                    access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "Bearer"
                    expires_in: 900
                    refresh_token: "new_rotated_refresh_token_string"
                    scope: "openid profile email"
                device_code:
                  summary: Device Code Exchange
                  value:
                    access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "Bearer"
                    expires_in: 900
                    refresh_token: "opaque_refresh_token_string"
                    scope: "openid profile email"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TokenRequest:
      type: object
      required:
        - grant_type
      properties:
        grant_type:
          type: string
          enum:
            - authorization_code
            - refresh_token
            - urn:ietf:params:oauth:grant-type:device_code
        code:
          type: string
          description: Authorization code (for authorization_code grant)
        redirect_uri:
          type: string
          description: Redirect URI (for authorization_code grant)
        code_verifier:
          type: string
          description: PKCE code verifier (for authorization_code grant)
        refresh_token:
          type: string
          description: Refresh token (for refresh_token grant)
        device_code:
          type: string
          description: Device code (for device_code grant)
        client_id:
          type: string
          description: Client ID
        client_secret:
          type: string
          description: Client secret (for confidential clients)
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token lifetime in seconds
        refresh_token:
          type: string
          description: Opaque refresh token
        id_token:
          type: string
          description: JWT ID token
        scope:
          type: string
          description: Granted scopes
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
            - server_error
            - temporarily_unavailable
            - authorization_pending
            - slow_down
            - expired_token
        error_description:
          type: string
        error_uri:
          type: string
          format: uri
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "invalid_request"
            error_description: "Missing required parameter: client_id"
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "invalid_grant"
            error_description: "Invalid authorization code"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token